refactor: move bio into dedicated `ProfileBio` component
